Backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
python-dotenv==1.0.1
firebase-admin==6.5.0
boto3==1.38.41
google-generativeai==0.6.0
pytz==2025.1
PyMuPDF==1.23.26
python-docx==1.1.2
python-multipart==0.0.12
aiofiles==24.1.0
pydantic==2.10.4
httpx==0.28.1
